[shardformer] simplify execute_w_pass_grad_accum & execute_w_pass
